♻️ refactor(c): split serialize headers, tidy layout
serialize/internal.h had accreted three concerns behind one include: the
growable code-point buffer with its reserve/append hot path, the escape
and tag-writer formatters, and the block predicates the markdown and
plain-text renderers share. The single 860-line header made the buffer
primitive hard to find and obscured what each output mode depends on.

The buffer moves to serialize/buffer.h and the shared block predicates to
serialize/markdown_internal.h. internal.h includes both at the positions
their definitions used to occupy, so every translation unit that pulls
internal.h sees the same definitions in the same order as before. The
split stays purely physical: no .c include changes, and each serialize
unit preprocesses byte-for-byte to its pre-split form, so the compiler
emits identical code and the CodSpeed instruction-count gate is
unaffected. Moving a static inline between headers can otherwise shift
the LTO inliner's decisions, which this layout avoids by keeping every
unit's include expansion unchanged.

Three smaller tidy-ups ride along: query/find/find.c loses its
single-file directory and becomes query/find.c, docs/reference.rst no
longer claims eight namespaces for a fifteen-page toctree (and now names
the validate and conformance pages it had skipped), and a _c/README.md
records how each subsystem lines up across the C tree, the Python module,
the tests directory, and the reference page.
