+++ docs/reference.rst
@@ -5,7 +5,7 @@
 The complete public API, grouped by task. Types and signatures come from the ``turbohtml._html`` extension's stub; the
 prose is the C docstrings, so the reference cannot drift from the compiled core.
 
-The pages follow the eight namespaces, in the order the :doc:`how-to guides </how-to/index>` and :doc:`migration guides
+The pages are grouped by task, in the order the :doc:`how-to guides </how-to/index>` and :doc:`migration guides
 </migration/index>` use. :doc:`reference/parsing` turns markup into a tree, :doc:`reference/nodes` is the node model and
 the navigation, query, and mutation methods shared by every node, and :doc:`reference/tokenizer` is the low-level token
 surface. :doc:`reference/detect` sniffs the character encoding of raw bytes. :doc:`reference/query` is the CSS, XPath,
@@ -14,7 +14,9 @@ and soupsieve-shaped search surface, and :doc:`reference/cssom` resolves the CSS
 translates CSS selectors to XPath 1.0, and :doc:`reference/transform` applies an XSLT 1.0 stylesheet.
 :doc:`reference/extract` and :doc:`reference/structured-data` pull the article, its metadata, and the JSON-LD /
 Microdata / OpenGraph records out of a page. :doc:`reference/build` constructs a tree in code, and
-:doc:`reference/serialize` renders one back to HTML, Markdown, or plain text.
+:doc:`reference/serialize` renders one back to HTML, Markdown, or plain text. :doc:`reference/validate` checks a
+document against an XSD 1.0 or RELAX NG schema, and :doc:`reference/conformance` runs the HTML5 authoring-conformance
+rules.
 
 .. currentmodule:: turbohtml
