+++ src/turbohtml/_c/serialize/internal.h
@@ -1,9 +1,11 @@
 /* Serialization primitives shared by every output mode (#document, HTML,
-   minify, markdown, text, readability). The growable code-point buffer, the
-   escape formatters, the start/close-tag writers, the attribute ordering and
-   the markdown block predicates live here as `static inline` so each
-   serialize translation unit keeps its own inlined copy of the hot append and
-   escape loops with no cross-TU call. */
+   minify, markdown, text, readability). The escape formatters, the
+   start/close-tag writers and the attribute ordering live here as
+   `static inline`; the growable buffer they fill is included from buffer.h and
+   the markdown/plain-text block predicates from markdown_internal.h, at the
+   positions their definitions used to occupy, so a translation unit that
+   includes this header sees the same definitions in the same order as before
+   the split. */
 
 #ifndef TURBOHTML_SERIALIZE_INTERNAL_H
 #define TURBOHTML_SERIALIZE_INTERNAL_H
@@ -18,111 +20,7 @@
 
 #include <string.h>
 
-typedef struct {
-    Py_UCS4 *data;
-    Py_ssize_t len;
-    Py_ssize_t cap;
-    int failed;
-} sbuf;
-
-/* Grow the buffer so at least extra more code points fit, doubling so a run of
-   appends stays amortized O(1). */
-static inline void sbuf_reserve(sbuf *out, Py_ssize_t extra) {
-    if (out->len + extra <= out->cap) {
-        return;
-    }
-    size_t cap;
-    size_t bytes;
-    int grew = th_grow_cap((size_t)(out->len + extra), (size_t)out->cap, 256, sizeof(Py_UCS4), &cap, &bytes);
-    if (!grew) {         /* GCOVR_EXCL_BR_LINE: size overflow needs a length no allocation could hold */
-        out->failed = 1; /* GCOVR_EXCL_LINE: size-overflow path, unreachable from a test */
-        return;          /* GCOVR_EXCL_LINE: size-overflow path, unreachable from a test */
-    }
-    Py_UCS4 *grown = PyMem_Realloc(out->data, bytes);
-    if (grown == NULL) { /* GCOVR_EXCL_BR_LINE: allocation failure cannot be forced from a test */
-        out->failed = 1; /* GCOVR_EXCL_LINE: allocation-failure path, unreachable from a test */
-        return;          /* GCOVR_EXCL_LINE: allocation-failure path, unreachable from a test */
-    }
-    out->data = grown;
-    out->cap = (Py_ssize_t)cap;
-}
-
-static inline void sbuf_putc(sbuf *out, Py_UCS4 character) {
-    sbuf_reserve(out, 1);
-    if (out->failed) { /* GCOVR_EXCL_BR_LINE: allocation failure cannot be forced from a test */
-        return;        /* GCOVR_EXCL_LINE: allocation-failure path, unreachable from a test */
-    }
-    out->data[out->len++] = character;
-}
-
-/* Append a run of code points in one bulk copy after a single capacity check. */
-static inline void sbuf_put_run(sbuf *out, const Py_UCS4 *text, Py_ssize_t len) {
-    if (len == 0) { /* an empty run carries a NULL text pointer, and memcpy declares its source non-null even for 0 */
-        return;
-    }
-    sbuf_reserve(out, len);
-    if (out->failed) { /* GCOVR_EXCL_BR_LINE: allocation failure cannot be forced from a test */
-        return;        /* GCOVR_EXCL_LINE: allocation-failure path, unreachable from a test */
-    }
-    memcpy(out->data + out->len, text, (size_t)len * sizeof(Py_UCS4));
-    out->len += len;
-}
-
-static inline void sbuf_puts(sbuf *out, const char *str) {
-    Py_ssize_t len = (Py_ssize_t)strlen(str);
-    sbuf_reserve(out, len);
-    if (out->failed) { /* GCOVR_EXCL_BR_LINE: allocation failure cannot be forced from a test */
-        return;        /* GCOVR_EXCL_LINE: allocation-failure path, unreachable from a test */
-    }
-    for (Py_ssize_t index = 0; index < len; index++) {
-        out->data[out->len + index] = (Py_UCS4)(unsigned char)str[index];
-    }
-    out->len += len;
-}
-
-static inline void sbuf_put_ucs4(sbuf *out, const Py_UCS4 *text, Py_ssize_t len) {
-    sbuf_put_run(out, text, len);
-}
-
-/* Write well-formed UTF-8 bytes (an interned attribute name) as code points. The
-   common all-ASCII name is copied in bulk; only a name with a byte >= 0x80
-   (a foreign mixed-case attribute) takes the per-code-point decoder. */
-static inline void sbuf_put_utf8(sbuf *out, const char *bytes, Py_ssize_t len) {
-    Py_ssize_t index = 0;
-    while (index < len && (unsigned char)bytes[index] < 0x80) {
-        index++;
-    }
-    if (index > 0) {
-        sbuf_reserve(out, index);
-        if (out->failed) { /* GCOVR_EXCL_BR_LINE: allocation failure cannot be forced from a test */
-            return;        /* GCOVR_EXCL_LINE: allocation-failure path */
-        }
-        for (Py_ssize_t ascii = 0; ascii < index; ascii++) {
-            out->data[out->len + ascii] = (Py_UCS4)(unsigned char)bytes[ascii];
-        }
-        out->len += index;
-    }
-    while (index < len) {
-        unsigned char lead = (unsigned char)bytes[index];
-        Py_UCS4 character;
-        if (lead < 0x80) {
-            character = lead;
-            index += 1;
-        } else if (lead < 0xE0) {
-            character = (Py_UCS4)(lead & 0x1F) << 6 | ((unsigned char)bytes[index + 1] & 0x3F);
-            index += 2;
-        } else if (lead < 0xF0) {
-            character = (Py_UCS4)(lead & 0x0F) << 12 | ((unsigned char)(bytes[index + 1] & 0x3F)) << 6 |
-                        ((unsigned char)bytes[index + 2] & 0x3F);
-            index += 3;
-        } else {
-            character = (Py_UCS4)(lead & 0x07) << 18 | ((unsigned char)(bytes[index + 1] & 0x3F)) << 12 |
-                        ((unsigned char)(bytes[index + 2] & 0x3F)) << 6 | ((unsigned char)bytes[index + 3] & 0x3F);
-            index += 4;
-        }
-        sbuf_putc(out, character);
-    }
-}
+#include "serialize/buffer.h" /* the sbuf struct and its reserve/append hot path */
 
 /* The #document attribute sort works in fixed stack buffers: at most this many
    attributes are ordered, each rendered name capped at this many bytes. */
@@ -745,117 +643,6 @@ static inline int foreign_attr_namespaced(const char *lower, Py_ssize_t len) {
     return 0;
 }
 
-/* A block-level element opens its own line(s); everything else is inline and
-   flows into the surrounding line. Shared by the markdown and text renderers. */
-static inline int is_md_block(uint16_t atom) {
-    switch (atom) {
-    case TH_TAG_HTML:
-    case TH_TAG_BODY:
-    case TH_TAG_P:
-    case TH_TAG_DIV:
-    case TH_TAG_SECTION:
-    case TH_TAG_ARTICLE:
-    case TH_TAG_HEADER:
-    case TH_TAG_FOOTER:
-    case TH_TAG_NAV:
-    case TH_TAG_ASIDE:
-    case TH_TAG_MAIN:
-    case TH_TAG_FIGURE:
-    case TH_TAG_FIGCAPTION:
-    case TH_TAG_ADDRESS:
-    case TH_TAG_BLOCKQUOTE:
-    case TH_TAG_PRE:
-    case TH_TAG_HR:
-    case TH_TAG_H1:
-    case TH_TAG_H2:
-    case TH_TAG_H3:
-    case TH_TAG_H4:
-    case TH_TAG_H5:
-    case TH_TAG_H6:
-    case TH_TAG_UL:
-    case TH_TAG_OL:
-    case TH_TAG_LI:
-    case TH_TAG_DL:
-    case TH_TAG_DT:
-    case TH_TAG_DD:
-    case TH_TAG_MENU:
-    case TH_TAG_DETAILS:
-    case TH_TAG_SUMMARY:
-    case TH_TAG_TABLE:
-        return 1;
-    default:
-        return 0;
-    }
-}
-
-/* Elements whose entire subtree contributes nothing to a text/markdown rendering:
-   document metadata and scripts. <script>/<style> hold code, not prose, in every
-   namespace (an inline-SVG stylesheet is CSS just like an HTML one), so they are
-   dropped regardless of ns; <head> is an HTML-only concept. */
-static inline int is_md_skipped(const th_node *node) {
-    if (node->atom == TH_TAG_SCRIPT || node->atom == TH_TAG_STYLE) {
-        return 1;
-    }
-    return node->ns == TH_NS_HTML && node->atom == TH_TAG_HEAD;
-}
-
-static inline Py_ssize_t md_put_decimal(sbuf *out, Py_ssize_t number) {
-    Py_UCS4 digits[20];
-    Py_ssize_t count = 0;
-    do {
-        digits[count++] = (Py_UCS4)('0' + (int)(number % 10));
-        number /= 10;
-    } while (number > 0);
-    for (Py_ssize_t index = count - 1; index >= 0; index--) {
-        sbuf_putc(out, digits[index]);
-    }
-    return count;
-}
-
-/* The value of one attribute by interned name, or NULL with *len 0 when the
-   attribute is absent or valueless. */
-static inline const Py_UCS4 *md_attr(th_tree *tree, th_node *node, const char *name, Py_ssize_t *len) {
-    Py_ssize_t index = th_node_attr_find(tree, node, name, (Py_ssize_t)strlen(name));
-    if (index < 0 || node->attrs[index].value == NULL) {
-        *len = 0;
-        return NULL;
-    }
-    *len = node->attrs[index].value_len;
-    return node->attrs[index].value;
-}
-
-static inline Py_ssize_t md_row_cells(th_node *row) {
-    Py_ssize_t count = 0;
-    for (th_node *cell = row->first_child; cell != NULL; cell = cell->next_sibling) {
-        if (cell->type == TH_NODE_ELEMENT && (cell->atom == TH_TAG_TD || cell->atom == TH_TAG_TH)) {
-            count++;
-        }
-    }
-    return count;
-}
-
-/* Trim the buffer in place under the document-strip mode, returning how many
-   leading code points were removed so a caller can shift recorded offsets. */
-static inline Py_ssize_t md_trim(sbuf *out, int mode) {
-    Py_ssize_t start = 0;
-    if (mode == TH_MD_DOC_STRIP || mode == TH_MD_DOC_LSTRIP) {
-        /* strip leading blank lines only; block content never starts with a space
-           except an indented code block, whose indent must survive */
-        while (start < out->len && out->data[start] == '\n') {
-            start++;
-        }
-    }
-    Py_ssize_t end = out->len;
-    if (mode == TH_MD_DOC_STRIP || mode == TH_MD_DOC_RSTRIP) {
-        while (end > start && is_space(out->data[end - 1])) {
-            end--;
-        }
-    }
-    if (start > 0) {
-        memmove(out->data, out->data + start, (size_t)(end - start) * sizeof(Py_UCS4));
-    }
-    out->len = end - start;
-    return start; /* how many leading code points were removed, to shift annotations */
-}
+#include "serialize/markdown_internal.h" /* the markdown/plain-text block predicates */
 
 #endif /* TURBOHTML_SERIALIZE_INTERNAL_H */
